requirements.txt
langchain==0.3.26
openai==1.37.0

redis==5.0.4
neo4j==5.19.0
sentence-transformers==2.6.1

transformers==4.41.1
torch==2.2.0
accelerate>=0.27.0
googlesearch-python
networkx==3.2.1
graphviz==0.20.1
python-dotenv==1.0.0
langchain_community
chromadb>=1.0.9
langchain-chroma==0.2.4